turn off driver selector for now, created duo GameStart.cpp/.h, game is initiated in GameEngine's main
