+++ Warzone/GameEngine.cpp
@@ -1,142 +1,10 @@
 #include "GameEngine.h"
-#include "Map.h"
+#include "GameStart.h"
 #include <iostream>
-#include <filesystem>
 using namespace std;
-namespace fs = std::filesystem;
 
-const int minPlayer = 2;
-const int maxPlayer = 5;
-
-// if input is not of type integer
-void cinFail() {
-	if (cin.fail())
-	{
-		cin.clear();
-		cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
-	}
-}
-
-void GameEngine::GameStart()
-{
-	cout << "********** GAME START **********" << endl;
-	this->GameStart_readMapDirectory();
-	this->GameStart_loadMaps();
-	this->GameStart_storeMaps();
-	this->GameStart_validatingMaps();
-	this->GameStart_promptUserToSelectMap();
-	this->GameStart_promptUserToSelectNumberOfPlayers();
-}
-
-void GameEngine::GameStart_readMapDirectory()
-{
-	cout << "Reading map directory..." << endl;
-	int i = 0;
-	for (const auto& p : fs::directory_iterator("maps")) {
-		const auto path = p.path();
-		if (p.is_regular_file() && path.filename().extension() == ".map") {
-			i++;
-			cout << "> map " << i << ": maps/" + path.filename().string() << endl; // e.g. "maps/canada.map"
-			mapFiles.push_back("maps/" + path.filename().string());
-		}
-	}
-	cout << "done..." << endl;
-}
-
-void GameEngine::GameStart_loadMaps()
-{
-	cout << "Loading maps..." << endl;
-	list<string>::iterator it;
-	for (it = mapFiles.begin(); it != mapFiles.end(); ++it) {
-		string filename = (*it);
-		mapLoaders.push_back(new MapLoader((*it)));
-	}
-	cout << "done..." << endl;
-}
-
-void GameEngine::GameStart_storeMaps()
-{
-	cout << "Storing map..." << endl;
-	vector<MapLoader*>::iterator iter;
-	for (iter = mapLoaders.begin(); iter != mapLoaders.end(); ++iter) {
-		(*iter)->storeContinents();
-		(*iter)->storeTerritories();
-		(*iter)->storeTerritoriesWithBorders();
-	}
-	cout << "done..." << endl;
-}
-
-void GameEngine::GameStart_validatingMaps()
-{
-	cout << "Validating maps..." << endl;
-	int invalidMapIndex = NULL;
-	for (int i = 0; i < mapLoaders.size(); i++) {
-		Map* map;
-		map = new Map();
-		map->createMap(mapLoaders.at(i)->getTerritories().size() + 1, true);
-		for (int j = 0; j < mapLoaders.at(i)->getTerritoriesWithBorders().size(); j++) {
-			map->addEdge(mapLoaders.at(i)->getTerritoriesWithBorders()[j]);
-		}
-		cout << "> validating " << mapLoaders.at(i)->getFileName() << endl;
-		map->validate();
-		if (map->getIsValidMapFile() == 0) {
-			invalidMapIndex = i; // store index of invalid map
-		}
-	}
-	cout << "done..." << endl;
-	if (invalidMapIndex != NULL) {
-		cout << "Gracefully rejecting " << mapLoaders.at(invalidMapIndex)->getFileName() << endl;
-		mapLoaders.erase(mapLoaders.begin() + invalidMapIndex);
-		cout << "done..." << endl;
-	}
-}
-
-void GameEngine::GameStart_promptUserToSelectNumberOfPlayers()
-{
-	int num;
-	cout << endl << "Enter the number players. Must be a value between " << minPlayer << "-" << maxPlayer << endl;
-	cin >> num;
-	while (num < minPlayer || num > maxPlayer)
-	{
-		cinFail();
-		cout << "Invalid input! Must be a value between " << minPlayer << "-" << maxPlayer << endl;
-		cout << "Please try again." << endl;
-		cin >> num;
-	}
-	this->GameStart_setNumPlayers(num);
-}
-
-void GameEngine::GameStart_promptUserToSelectMap()
-{
-	cout << endl << "Please choose a map..." << endl;
-	int chosenMapIndex;
-	for (unsigned k = 0; k < mapLoaders.size(); ++k) {
-		cout << "Enter number " << (k + 1) << " to choose " << mapLoaders.at(k)->getFileName() << endl;
-	}
-	cin >> chosenMapIndex;
-	while (chosenMapIndex <= 0 || chosenMapIndex > mapLoaders.size())
-	{
-		cinFail();
-		cout << "Invalid input! Must be a value between 1-" << mapLoaders.size() << endl;
-		cout << "Please enter a number again." << endl;
-		cin >> chosenMapIndex;
-	}
-	int chosenIndex = chosenMapIndex - 1;
-	cout << mapLoaders.at(chosenIndex)->getFileName() << " was chosen" << endl;
-	this->GameStart_setChosenMap(mapLoaders.at(chosenIndex)->getTerritoriesWithBorders());
-}
-
-void GameEngine::GameStart_setChosenMap(vector<Territory*> chosenMap)
-{
-	this->chosenMap = chosenMap;
-}
-
-void GameEngine::GameStart_setNumPlayers(int num)
-{
-	this->numPlayers = num;
-}
-
-int GameEngine::GameStart_getNumPlayers()
-{
-	return numPlayers;
+int main() {
+	GameStart g;
+	g.runAllFunctions();
+	return 0;
 }
