+++ Warzone/Warzone.vcxproj
@@ -142,6 +142,7 @@
   <ItemGroup>
     <ClInclude Include="DriverSelector.h" />
     <ClInclude Include="GameEngine.h" />
+    <ClInclude Include="GameStart.h" />
     <ClInclude Include="Map.h" />
     <ClInclude Include="MapLoader.h" />
     <ClInclude Include="Order.h" />
@@ -151,6 +152,7 @@
   <ItemGroup>
     <ClCompile Include="DriverSelector.cpp" />
     <ClCompile Include="GameEngine.cpp" />
+    <ClCompile Include="GameStart.cpp" />
     <ClCompile Include="Map.cpp" />
     <ClCompile Include="MapDriver.cpp" />
     <ClCompile Include="MapLoader.cpp" />
